+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 01. :gift: Added exception `dev.orne.http.client.HttpAuthenticationRequiredException`
 01. :gift: Added interface `dev.orne.http.client.AuthenticationOperation`
 01. :gift: Added exception `dev.orne.http.client.HttpAuthenticationFailedException`
+01. :gift: Added class `dev.orne.http.client.AbstractHttpServiceOperation`